Move PartidaService to arena-real
